Vampire & Roll Adjustments + Admin
This update adds a command that allows for pumping blood into physical attributes for vampires. Roll has been updated to use the temporary value if it exists on a sheet, which has been implemented for Abilities as well as Attributes in anticipation of boosts from Totem.

Additionally, there are two admin commands added: +approve and +unapprove. These are what they say on the tin, approving and unapproving characters.