hashtag shifter logic <3
This creates the basic formula for setting up types of shifters.
The shifter_bio.json file contains all of the basic data that shifter characters should have from tribes to breeds to aspects and so on.
The SetStats command has been updated to include logic that allows for setting pools based on splat and type.
Renown is also set based on type.
Additionally we now have all pools, bios, and powers for mages and changelings.